[fix] exhibitions on optional type

+++ HokutosaiApp/ExhibitionsViewController.swift
@@ -10,7 +10,7 @@ import UIKit
 
 class ExhibitionsViewController: UIViewController, UITableViewDelegate, UITableViewDataSource, LikeableTableViewCellDelegate, TabBarIntaractiveController {
 
-    private var exhibitions: [Exhibition]!
+    private var exhibitions: [Exhibition]?
     
     private var tableView: UITableView!
     private let cellIdentifier = "Exhibitions"
@@ -25,12 +25,12 @@ class ExhibitionsViewController: UIViewController, UITableViewDelegate, UITableV
         let loadingView = SimpleLoadingView(frame: self.view.frame)
         self.view.addSubview(loadingView)
         HokutosaiApi.GET(HokutosaiApi.Exhibitions.Exhibitions()) { response in
-            guard response.isSuccess else {
+            guard response.isSuccess, let data = response.model else {
                 loadingView.removeFromSuperview()
                 return
             }
             
-            self.exhibitions = response.model
+            self.exhibitions = data
             self.tableView.reloadData()
             loadingView.removeFromSuperview()
         }
@@ -57,8 +57,10 @@ class ExhibitionsViewController: UIViewController, UITableViewDelegate, UITableV
     }
     
     func tableView(tableView: UITableView, didSelectRowAtIndexPath indexPath: NSIndexPath) {
+        guard let exhibitions = self.exhibitions else { return }
+        
         let detailsView = StandardDetailsViewController()
-        detailsView.title = self.exhibitions[indexPath.row].title!
+        detailsView.title = exhibitions[indexPath.row].title!
         self.navigationController?.pushViewController(detailsView, animated: true)
         self.tableView.deselectRowAtIndexPath(indexPath, animated: true)
     }
@@ -86,14 +88,16 @@ class ExhibitionsViewController: UIViewController, UITableViewDelegate, UITableV
     func tableView(tableView: UITableView, cellForRowAtIndexPath indexPath: NSIndexPath) -> UITableViewCell {
         let cell = tableView.dequeueReusableCellWithIdentifier(cellIdentifier, forIndexPath: indexPath) as! StandardTableViewCell
         
-        cell.changeData(indexPath.row, data: self.exhibitions[indexPath.row])
+        cell.changeData(indexPath.row, data: self.exhibitions![indexPath.row])
         cell.delegate = self
         
         return cell
     }
     
     func like(index: Int, cell: LikeableTableViewCell) {
-        let exhibitionId = self.exhibitions[index].exhibitionId!
+        guard self.exhibitions != nil else { return }
+        
+        let exhibitionId = self.exhibitions![index].exhibitionId!
         HokutosaiApi.POST(HokutosaiApi.Exhibitions.Likes(exhibitionId: exhibitionId)) { response in
             guard let result = response.model else {
                 self.presentViewController(ErrorAlert.Server.failureSendRequest(), animated: true, completion: nil)
@@ -101,14 +105,16 @@ class ExhibitionsViewController: UIViewController, UITableViewDelegate, UITableV
                 return
             }
             
-            self.exhibitions[index].liked = result.liked
-            self.exhibitions[index].likesCount = result.likesCount
+            self.exhibitions![index].liked = result.liked
+            self.exhibitions![index].likesCount = result.likesCount
             cell.updateLikes(exhibitionId)
         }
     }
     
     func dislike(index: Int, cell: LikeableTableViewCell) {
-        let exhibitionId = self.exhibitions[index].exhibitionId!
+        guard self.exhibitions != nil else { return }
+        
+        let exhibitionId = self.exhibitions![index].exhibitionId!
         HokutosaiApi.DELETE(HokutosaiApi.Exhibitions.Likes(exhibitionId: exhibitionId)) { response in
             guard let result = response.model else {
                 self.presentViewController(ErrorAlert.Server.failureSendRequest(), animated: true, completion: nil)
@@ -116,8 +122,8 @@ class ExhibitionsViewController: UIViewController, UITableViewDelegate, UITableV
                 return
             }
             
-            self.exhibitions[index].liked = result.liked
-            self.exhibitions[index].likesCount = result.likesCount
+            self.exhibitions![index].liked = result.liked
+            self.exhibitions![index].likesCount = result.likesCount
             cell.updateLikes(exhibitionId)
         }
     }
